+++ src/ZenExtensions.Slack.Models/Message.cs
@@ -49,14 +49,14 @@ namespace ZenExtensions.Slack.Models
         /// <param name="attachment">Instance of <see cref="Attachment"/></param>
         /// <returns>Instance of current <see cref="Message"/></returns>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="attachment"/> is null</exception>
-        /// <exception cref="InvalidOperationException">Thrown when there are already 10 attachments attached with message</exception>
+        /// <exception cref="ArgumentException">Thrown when there are already 10 attachments attached with message</exception>
         public Message AddAttachment([NotNull] Attachment attachment)
         {
             ArgumentNullException.ThrowIfNull(attachment, nameof(attachment));
             Attachments ??= new List<Attachment>();
             if (Attachments.Count >= 10)
             {
-                throw new InvalidOperationException("Cannot add more than 10 attachments to a message");
+                throw new ArgumentException("Cannot add more than 10 attachments to a message");
             }
             Attachments.Add(attachment);
             return this;